kube-apiserver: split setup code
